// src/app/modules/contact/core/models/contact.enum.ts
export enum EContactStatus {
  PENDING = "pending",
  PROCESS = "process",
  COMPLETED = "completed",
}

export enum EContactStatusL {
  pending = "Beklemede",
  process = "İşlemde",
  completed = "Tamamlandı",
}

export enum EContactAdminType {
  NORMAL = "normal",
  SUPER = "super",
}

export enum EContactAdminTypeL {
  normal = "Normal Yetkili",
  super = "Süper Yetkili",
}